pyproject.toml
[project]
name = "virelix"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=46.0.6",
    "django>=6.0.3",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.6",
    "psycopg2>=2.9.11",
]
